gmtColormap.py: change np.float to float
Addresses:
https://github.com/dshean/imview/issues/4
https://github.com/dshean/demcoreg/issues/57

+++ imview/lib/gmtColormap.py
@@ -71,10 +71,10 @@ def gmtColormap(fileName,GMTPath = None, reverse=False):
       b.append(btemp)
 
       nTable = len(r)
-      x = np.array( x , np.float)
-      r = np.array( r , np.float)
-      g = np.array( g , np.float)
-      b = np.array( b , np.float)
+      x = np.array(x, float)
+      r = np.array(r, float)
+      g = np.array(g, float)
+      b = np.array(b, float)
 
       if reverse:
         r = r[::-1]
